Move multiprocessing into ocean_basin_proximity.py.
So it's only in one place (not also in 01_generate_distance_grids.py).
